requirements.txt
djangorestframework
requests
python-dotenv